requirements.txt
slack-bolt==1.18.0
slack-sdk==3.26.1
jira==3.5.2
pandas==2.1.4
matplotlib==3.8.2
reportlab==4.0.8
python-dotenv==1.0.0
requests==2.31.0 